// libs/explorer-lib/src/lib/constants/enums/explorer-actions.enum.ts
/*
 * Copyright (c) 2023-2024. IdrisovII & EustroSoft.org
 *
 * This file is part of eustrosoft-front project.
 * See the LICENSE file at the project root for licensing information.
 */

export enum ExplorerRequestActions {
  VIEW = 'view',
  CREATE = 'create',
  UPLOAD_CHUNKS_BASE64 = 'upload_chunks',
  UPLOAD_CHUNKS_BINARY = 'upload_chunks_binary',
  UPLOAD_CHUNKS_HEX = 'upload_chunks_hex',
  MOVE = 'move',
  COPY = 'copy',
  DELETE = 'delete',
  TICKET = 'ticket',
}